aggiunta classi

+++ classes/product.php
@@ -7,7 +7,7 @@ class product {
     public $quality;
     public $animalType;
 
-    public function __construct(string $_name, int $_price, string $_quality, string $_animalType){
+    public function __construct(string $_name, float $_price, string $_quality, string $_animalType){
         $this->name = $_name;
         $this->price = $_price;
         $this->quality = $_quality;
